Add GroupMixin to ChannelMutator (temporarily)
