+++ mmrazor/models/mutators/channel_mutator/one_shot_channel_mutator.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from typing import Dict, Type, Union
+from typing import Dict, Type, Union, Any
 
 from mmrazor.models.mutables import OneShotMutableChannelUnit
 from mmrazor.registry import MODELS
@@ -28,14 +28,18 @@ class OneShotChannelMutator(ChannelMutator[OneShotMutableChannelUnit]):
 
     def min_choices(self) -> Dict:
         """Return the minimal pruning subnet(structure)."""
-        template = self.choice_template
-        for key in template:
-            template[key] = self._name2unit[key].min_choice
-        return template
+        min_choices = dict()
+        for group_id, modules in self.search_groups.items():
+            min_choices[group_id] = modules[0].min_choice
+
+        return min_choices
 
     def max_choices(self) -> Dict:
         """Return the maximal pruning subnet(structure)."""
-        template = self.choice_template
-        for key in template:
-            template[key] = self._name2unit[key].max_choice
-        return template
+        max_choices = dict()
+        for group_id, modules in self.search_groups.items():
+            max_choices[group_id] = modules[0].max_choice
+
+        return max_choices
+
+
