+++ mmrazor/models/mutators/channel_mutator/slimmable_channel_mutator.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from typing import Dict, List, Optional
+from typing import Any, Dict, List, Optional
 
 from mmrazor.models.mutables import SlimmableChannelUnit
 from mmrazor.registry import MODELS
@@ -31,6 +31,16 @@ class SlimmableChannelMutator(ChannelMutator[SlimmableChannelUnit]):
         super().__init__(channel_unit_cfg, parse_cfg, init_cfg)
 
         self.subnets = self._prepare_subnets(self.units_cfg)
+
+    def set_choices(self, config: Dict[str, float]):
+        """Set choices."""
+        for name, choice in config.items():
+            unit = self._name2unit[name]
+            unit.current_choice = choice
+
+    def sample_choices(self) -> Dict[str, float]:
+        """Sample choices(pruning structure)."""
+        raise RuntimeError
 
     # private methods
