+++ tests/test_models/test_algorithms/test_prune_algorithm.py
@@ -114,7 +114,8 @@ class TestItePruneAlgorithm(unittest.TestCase):
         model = MODELS.build(MODEL_CFG)
         mutator = MODELS.build(MUTATOR_CONFIG_FLOAT)
         mutator.prepare_from_supernet(model)
-        prune_target = mutator.sample_choices()
+        mutator.set_choices(mutator.sample_choices())
+        prune_target = mutator.choice_template
 
         epoch = 10
         epoch_step = 2
@@ -135,9 +136,11 @@ class TestItePruneAlgorithm(unittest.TestCase):
                     data['inputs'], data['data_samples'], mode='loss')
 
         current_choices = algorithm.mutator.current_choices
+        group_prune_target = algorithm.group_target_pruning_ratio(
+            prune_target, mutator.search_groups)
         for key in current_choices:
             self.assertAlmostEqual(
-                current_choices[key], prune_target[key], delta=0.1)
+                current_choices[key], group_prune_target[key], delta=0.1)
 
     def test_load_pretrained(self):
         epoch_step = 2
@@ -158,7 +161,7 @@ class TestItePruneAlgorithm(unittest.TestCase):
         algorithm = ItePruneAlgorithm(
             model_cfg,
             mutator_cfg=MUTATOR_CONFIG_NUM,
-            target_pruning_ratio={},
+            target_pruning_ratio=None,
             step_epoch=epoch_step,
             prune_times=times,
         ).to(DEVICE)
